scripts/cloud_functions/requirements.txt
google-cloud-language==2.14.0
google-cloud-bigquery==3.12.0
google-cloud-logging==3.5.0